+++ internal/middleware/cors.go
@@ -1,0 +1,20 @@
+package middleware
+
+import (
+	"net/http"
+)
+
+func Cors() func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			// TODO: fix to allow only "www.serj-tubin.com"
+
+			//Allow CORS here By * or specific origin
+			w.Header().Set("Access-Control-Allow-Origin", "*")
+			w.Header().Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
+			w.Header().Set("Access-Control-Allow-Headers", "Accept, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization")
+
+			next.ServeHTTP(w, r)
+		})
+	}
+}
